requirements.txt
Django==4.2.6
djangorestframework==3.14.0
celery[redis]==5.3.4
redis==4.6.0
gunicorn==21.2.0
django-celery-results==2.5.1
mysqlclient==2.2.0
flower==2.0.1
